requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
fastapi==0.75.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
Pillow==8.4.0
pydantic==1.9.0
python-magic==0.4.25
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.16.0
zipp==3.6.0
